requirements.txt
Pillow==9.2.0
Pillow==10.3.0
streamlit==1.12.0
